requirements.txt
django
pandas
numpy
scikit-surprise==1.1.1